test/e2e: move package var declarations and bootstrap funcs out of _test.go files
Go excludes _test.go files from normal package imports, so all the
work exporting SystemTestInput/globals/etc in prior commits was
necessary but not sufficient - the var declarations and setup
functions themselves lived in _test.go files and were simply
undefined for an external `import e2e ...` consumer.

Moves (pure relocation, no logic changes):
- Package vars Ctx/TestClient/RestConfig/Clientset/Conf/CleanupFlags/
  CleanupTracker/ArtifactsDir/CollectResources/input/VPCOVNInput and
  the test-flag-only externalTest var: e2e_suite_test.go and
  system_test.go -> new globals.go
- SetInput/SystemSetupBeforeSuite/createNGCImagePullSecret/
  AnnotateAndLabelNodes/GetProvisionDPUClustersInput/validateFlags:
  system_test.go -> new system_bootstrap.go
- ProvisioningBeforeSuite: provisioning_test.go -> provisioning.go
- SNAPBeforeSuite: snap_test.go -> new snap.go
- SDNBeforeSuite: sdn_test.go -> new sdn.go
- VPCOVNBeforeSuite: vpcovn_test.go -> vpcovn.go
- CreateDPUWorkerNodes: scale_test.go -> new scale.go

_test.go files keep only their Describe/It/BeforeSuite/AfterSuite
Ginkgo registrations and the TestE2E entrypoint, which only matter
when running via 'go test' - confirmed both 'go build ./...' (the
external-import path) and 'go test ./test/e2e/... -list .*' (the
internal go-test path) pass after this change.

    gitleaks.version: 8.30.0
    gitleaks.check-secrets: ENABLED
