+++ test/e2e/snap.go
@@ -20,16 +20,6 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 )
 
-//nolint:dupl
-var _ = Describe("DPF SNAP tests", Labels{Domain.DPFSystem, Domain.SNAP}, func() {
-
-	BeforeEach(func() {
-		By("Placeholder for the SNAP BeforeEach test condition check/action")
-	})
-	AfterEach(func() {
-		By("Placeholder for the SNAP AfterEach test condition check/action/cleanup")
-	})
-
-	Context("Placeholder for the SNAP tests", Labels{Domain.RequiresNodes}, Serial, func() {
-	})
-})
+func SNAPBeforeSuite() {
+	By("Setting SNAP configs for the test")
+}
